feat(time-picker): add storybook controls
