--- packages/react/src/components/TimePicker/TimePicker-story.js
@@ -1,94 +1,0 @@
-/**
- * Copyright IBM Corp. 2016, 2018
- *
- * This source code is licensed under the Apache-2.0 license found in the
- * LICENSE file in the root directory of this source tree.
- */
-
-import React from 'react';
-
-import {
-  withKnobs,
-  boolean,
-  number,
-  select,
-  text,
-} from '@storybook/addon-knobs';
-import TimePicker from '../TimePicker';
-import TimePickerSelect from '../TimePickerSelect';
-import SelectItem from '../SelectItem';
-import mdx from './TimePicker.mdx';
-
-const sizes = {
-  'Small  (sm)': 'sm',
-  'Medium (md) - default': undefined,
-  'Large  (lg)': 'lg',
-};
-
-const props = {
-  timepicker: () => ({
-    pattern: text(
-      'Regular expression for the value (pattern in <TimePicker>)',
-      '(1[012]|[1-9]):[0-5][0-9](\\s)?'
-    ),
-    placeholder: text(
-      'Placeholder text (placeholder in <TimePicker>)',
-      'hh:mm'
-    ),
-    disabled: boolean('Disabled (disabled in <TimePicker>)', false),
-    light: boolean('Light variant (light in <TimePicker>)', false),
-    labelText: text('Label text (labelText in <TimePicker>)', 'Select a time'),
-    invalid: boolean(
-      'Show form validation UI (invalid in <TimePicker>)',
-      false
-    ),
-    invalidText: text(
-      'Form validation UI content (invalidText in <TimePicker>)',
-      'A valid value is required'
-    ),
-    maxLength: number('Maximum length (maxLength in <TimePicker>)', 5),
-    size: select('Field size (size)', sizes, undefined) || undefined,
-  }),
-  select: () => ({
-    disabled: boolean('Disabled (disabled in <TimePickerSelect>)', false),
-    labelText: text(
-      'Label text (labelText in <TimePickerSelect>)',
-      'Please select'
-    ),
-    ['aria-label']: text(
-      'ARIA label (aria-label in <TimePickerSelect>)',
-      'open list of options'
-    ),
-  }),
-};
-
-export default {
-  title: 'Components/TimePicker',
-  component: TimePicker,
-  decorators: [withKnobs],
-  subcomponents: {
-    TimePickerSelect,
-    SelectItem,
-  },
-  parameters: {
-    docs: {
-      page: mdx,
-    },
-  },
-};
-
-export const Default = () => {
-  const selectProps = props.select();
-  return (
-    <TimePicker id="time-picker" {...props.timepicker()}>
-      <TimePickerSelect id="time-picker-select-1" {...selectProps}>
-        <SelectItem value="AM" text="AM" />
-        <SelectItem value="PM" text="PM" />
-      </TimePickerSelect>
-      <TimePickerSelect id="time-picker-select-2" {...selectProps}>
-        <SelectItem value="Time zone 1" text="Time zone 1" />
-        <SelectItem value="Time zone 2" text="Time zone 2" />
-      </TimePickerSelect>
-    </TimePicker>
-  );
-};
