+++ packages/react/src/components/TimePicker/TimePicker.js
@@ -6,256 +6,224 @@
  */
 
 import PropTypes from 'prop-types';
-import React, { Component } from 'react';
-import classNames from 'classnames';
-import { FeatureFlagContext } from '../FeatureFlags';
-import { PrefixContext } from '../../internal/usePrefix';
+import React from 'react';
+import cx from 'classnames';
+import { usePrefix } from '../../internal/usePrefix';
+import deprecate from '../../prop-types/deprecate';
 
-export default class TimePicker extends Component {
-  state = {};
+const TimePicker = React.forwardRef(function TimePicker(
+  {
+    children,
+    className,
+    disabled = false,
+    hideLabel,
+    id,
+    invalidText = 'Invalid time format.',
+    invalid = false,
+    labelText,
+    light = false,
+    maxLength = 5,
+    onChange = () => {},
+    onClick = () => {},
+    onBlur = () => {},
+    pattern = '(1[012]|[1-9]):[0-5][0-9](\\s)?',
+    placeholder = 'hh:mm',
+    size = 'md',
+    type = 'text',
+    value,
+    ...rest
+  },
+  ref
+) {
+  const prefix = usePrefix();
 
-  static propTypes = {
-    /**
-     * Pass in the children that will be rendered next to the form control
-     */
-    children: PropTypes.node,
+  const [isValue, setValue] = React.useState(value);
+  const [prevValue, setPrevValue] = React.useState(value);
 
-    /**
-     * Specify an optional className to be applied to the container node
-     */
-    className: PropTypes.string,
-
-    /**
-     * Specify whether the `<input>` should be disabled
-     */
-    disabled: PropTypes.bool,
-
-    /**
-     * Specify whether you want the underlying label to be visually hidden
-     */
-    hideLabel: PropTypes.bool,
-
-    /**
-     * Specify a custom `id` for the `<input>`
-     */
-    id: PropTypes.string.isRequired,
-
-    /**
-     * Specify whether the control is currently invalid
-     */
-    invalid: PropTypes.bool,
-
-    /**
-     * Provide the text that is displayed when the control is in an invalid state
-     */
-    invalidText: PropTypes.node,
-
-    /**
-     * Provide the text that will be read by a screen reader when visiting this
-     * control
-     */
-    labelText: PropTypes.node,
-
-    /**
-     * `true` to use the light version. For use on $ui-01 backgrounds only.
-     * Don't use this to make tile background color same as container background color.
-     */
-    light: PropTypes.bool,
-
-    /**
-     * Specify the maximum length of the time string in `<input>`
-     */
-    maxLength: PropTypes.number,
-
-    /**
-     * Optionally provide an `onBlur` handler that is called whenever the
-     * `<input>` loses focus
-     */
-    onBlur: PropTypes.func,
-
-    /**
-     * Optionally provide an `onChange` handler that is called whenever `<input>`
-     * is updated
-     */
-    onChange: PropTypes.func,
-
-    /**
-     * Optionally provide an `onClick` handler that is called whenever the
-     * `<input>` is clicked
-     */
-    onClick: PropTypes.func,
-
-    /**
-     * Specify the regular expression working as the pattern of the time string in `<input>`
-     */
-    pattern: PropTypes.string,
-
-    /**
-     * Specify the placeholder attribute for the `<input>`
-     */
-    placeholder: PropTypes.string,
-
-    /**
-     * Specify the size of the Time Picker.
-     */
-    size: PropTypes.oneOf(['sm', 'md', 'lg', 'xl']),
-
-    /**
-     * Specify the type of the `<input>`
-     */
-    type: PropTypes.string,
-
-    /**
-     * Specify the value of the `<input>`
-     */
-    value: PropTypes.string,
-  };
-
-  static defaultProps = {
-    type: 'text',
-    pattern: '(1[012]|[1-9]):[0-5][0-9](\\s)?',
-    placeholder: 'hh:mm',
-    maxLength: 5,
-    invalidText: 'Invalid time format.',
-    invalid: false,
-    disabled: false,
-    onChange: () => {},
-    onClick: () => {},
-    onBlur: () => {},
-    light: false,
-  };
-
-  static contextType = FeatureFlagContext;
-
-  static getDerivedStateFromProps({ value }, state) {
-    const { prevValue } = state;
-    return prevValue === value
-      ? null
-      : {
-          value,
-          prevValue: value,
-        };
+  if (value !== prevValue) {
+    setValue(value);
+    setPrevValue(value);
   }
 
-  render() {
-    const {
-      children,
-      className,
-      id,
-      labelText,
-      type,
-      pattern,
-      onChange,
-      onClick,
-      onBlur,
-      placeholder,
-      maxLength,
-      invalidText,
-      invalid,
-      hideLabel,
-      light,
-      size,
-      ...other
-    } = this.props;
-
-    const scope = this.context;
-    let enabled;
-
-    if (scope.enabled) {
-      enabled = scope.enabled('enable-v11-release');
+  function handleOnClick(evt) {
+    if (!disabled) {
+      setValue(isValue);
+      onClick(evt);
     }
-
-    return (
-      <PrefixContext.Consumer>
-        {(prefix) => {
-          const timePickerInputProps = {
-            className: classNames(
-              `${prefix}--time-picker__input-field`,
-              `${prefix}--text-input`,
-              [enabled ? null : className],
-              {
-                [`${prefix}--text-input--light`]: light,
-              }
-            ),
-
-            onChange: (evt) => {
-              if (!other.disabled) {
-                this.setState({
-                  value: evt.target.value,
-                });
-                onChange(evt);
-              }
-            },
-            onClick: (evt) => {
-              if (!other.disabled) {
-                this.setState({
-                  value: evt.target.value,
-                });
-                onClick(evt);
-              }
-            },
-            onBlur: (evt) => {
-              if (!other.disabled) {
-                this.setState({
-                  value: evt.target.value,
-                });
-                onBlur(evt);
-              }
-            },
-            pattern,
-            placeholder,
-            maxLength,
-            id,
-            type,
-            value: this.state.value,
-          };
-
-          const timePickerClasses = classNames({
-            [`${prefix}--time-picker`]: true,
-            [`${prefix}--time-picker--light`]: light,
-            [`${prefix}--time-picker--invalid`]: invalid,
-            [`${prefix}--time-picker--${size}`]: size,
-            [className]: className,
-          });
-
-          const labelClasses = classNames(`${prefix}--label`, {
-            [`${prefix}--visually-hidden`]: hideLabel,
-            [`${prefix}--label--disabled`]: other.disabled,
-          });
-
-          const label = labelText ? (
-            <label htmlFor={id} className={labelClasses}>
-              {labelText}
-            </label>
-          ) : null;
-
-          const error = invalid ? (
-            <div className={`${prefix}--form-requirement`}>{invalidText}</div>
-          ) : null;
-
-          return (
-            <div
-              className={
-                enabled
-                  ? classNames(`${prefix}--form-item`, className)
-                  : `${prefix}--form-item`
-              }>
-              {label}
-              <div className={timePickerClasses}>
-                <div className={`${prefix}--time-picker__input`}>
-                  <input
-                    {...other}
-                    {...timePickerInputProps}
-                    data-invalid={invalid ? invalid : undefined}
-                  />
-                </div>
-                {children}
-              </div>
-              {error}
-            </div>
-          );
-        }}
-      </PrefixContext.Consumer>
-    );
   }
-}
+
+  function handleOnChange(evt) {
+    if (!disabled) {
+      setValue(isValue);
+      onChange(evt);
+    }
+  }
+
+  function handleOnBlur(evt) {
+    if (!disabled) {
+      setValue(isValue);
+      onBlur(evt);
+    }
+  }
+
+  const timePickerInputClasses = cx(
+    `${prefix}--time-picker__input-field`,
+    `${prefix}--text-input`,
+    [className],
+    {
+      [`${prefix}--text-input--light`]: light,
+    }
+  );
+
+  const timePickerClasses = cx({
+    [`${prefix}--time-picker`]: true,
+    [`${prefix}--time-picker--light`]: light,
+    [`${prefix}--time-picker--invalid`]: invalid,
+    [`${prefix}--time-picker--${size}`]: size,
+    [className]: className,
+  });
+
+  const labelClasses = cx(`${prefix}--label`, {
+    [`${prefix}--visually-hidden`]: hideLabel,
+    [`${prefix}--label--disabled`]: disabled,
+  });
+
+  const label = labelText ? (
+    <label htmlFor={id} className={labelClasses}>
+      {labelText}
+    </label>
+  ) : null;
+
+  const error = invalid ? (
+    <div className={`${prefix}--form-requirement`}>{invalidText}</div>
+  ) : null;
+
+  return (
+    <div className={cx(`${prefix}--form-item`, className)}>
+      {label}
+      <div className={timePickerClasses}>
+        <div className={`${prefix}--time-picker__input`}>
+          <input
+            className={timePickerInputClasses}
+            data-invalid={invalid ? invalid : undefined}
+            disabled={disabled}
+            id={id}
+            maxLength={maxLength}
+            onClick={handleOnClick}
+            onChange={handleOnChange}
+            onBlur={handleOnBlur}
+            placeholder={placeholder}
+            pattern={pattern}
+            ref={ref}
+            type={type}
+            value={value}
+            {...rest}
+          />
+        </div>
+        {children}
+      </div>
+      {error}
+    </div>
+  );
+});
+
+TimePicker.propTypes = {
+  /**
+   * Pass in the children that will be rendered next to the form control
+   */
+  children: PropTypes.node,
+
+  /**
+   * Specify an optional className to be applied to the container node
+   */
+  className: PropTypes.string,
+
+  /**
+   * Specify whether the `<input>` should be disabled
+   */
+  disabled: PropTypes.bool,
+
+  /**
+   * Specify whether you want the underlying label to be visually hidden
+   */
+  hideLabel: PropTypes.bool,
+
+  /**
+   * Specify a custom `id` for the `<input>`
+   */
+  id: PropTypes.string.isRequired,
+
+  /**
+   * Specify whether the control is currently invalid
+   */
+  invalid: PropTypes.bool,
+
+  /**
+   * Provide the text that is displayed when the control is in an invalid state
+   */
+  invalidText: PropTypes.node,
+
+  /**
+   * Provide the text that will be read by a screen reader when visiting this
+   * control
+   */
+  labelText: PropTypes.node,
+
+  /**
+   * The `light` prop for `TimePicker` has been deprecated. It will be removed in v12. Use the `Layer` component instead.
+   */
+  light: deprecate(
+    PropTypes.bool,
+    'The `light` prop for `TimePicker` is no longer needed and has been deprecated. It will be removed in the next major release. Use the `Layer` component instead.'
+  ),
+
+  /**
+   * Specify the maximum length of the time string in `<input>`
+   */
+  maxLength: PropTypes.number,
+
+  /**
+   * Optionally provide an `onBlur` handler that is called whenever the
+   * `<input>` loses focus
+   */
+  onBlur: PropTypes.func,
+
+  /**
+   * Optionally provide an `onChange` handler that is called whenever `<input>`
+   * is updated
+   */
+  onChange: PropTypes.func,
+
+  /**
+   * Optionally provide an `onClick` handler that is called whenever the
+   * `<input>` is clicked
+   */
+  onClick: PropTypes.func,
+
+  /**
+   * Specify the regular expression working as the pattern of the time string in `<input>`
+   */
+  pattern: PropTypes.string,
+
+  /**
+   * Specify the placeholder attribute for the `<input>`
+   */
+  placeholder: PropTypes.string,
+
+  /**
+   * Specify the size of the Time Picker.
+   */
+  size: PropTypes.oneOf(['sm', 'md', 'lg']),
+
+  /**
+   * Specify the type of the `<input>`
+   */
+  type: PropTypes.string,
+
+  /**
+   * Specify the value of the `<input>`
+   */
+  value: PropTypes.string,
+};
+
+export default TimePicker;
